+++ CHANGELOG.md
@@ -22,7 +22,10 @@ adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 - checking device in zapsi database
 - checking logged user in zapsi database
 - creating order in zapsi database
+- creating idle in zapsi database
 - getting idles from zapsi database
+- checking user in zapsi database
+
 
 ### Changed
 - proper html formatting